extract test utils into a module, under feature
