+++ crates/compute/Cargo.toml
@@ -8,15 +8,21 @@ authors.workspace = true
 workspace = true
 
 [dependencies]
+binius_core = { path = "../core", default-features = false, optional = true }
 binius_field = { path = "../field", default-features = false }
 binius_ntt = { path = "../ntt", default-features = false }
 binius_math = { path = "../math", default-features = false }
 binius_utils = { path = "../utils", default-features = false }
 bytemuck.workspace = true
 itertools.workspace = true
+rand = { workspace = true, optional = true }
 thiserror.workspace = true
 
 [dev-dependencies]
 assert_matches.workspace = true
 binius_core = { path = "../core", default-features = false }
 rand = { workspace = true, features = ["std"] }
+
+[features]
+default = []
+test_utils = ["binius_core", "rand"]
