+++ crates/compute/src/lib.rs
@@ -10,3 +10,5 @@ pub mod alloc;
 pub mod cpu;
 pub mod layer;
 pub mod memory;
+#[cfg(any(test, feature = "test_utils"))]
+pub mod test_utils;
